Rename files and improve routes params
