+++ nextjs_frontend/src/app/negociacao/[chatID]/components/Message/message.tsx
