+++ nextjs_frontend/src/components/Header/components/NotificationCard/notification-card.tsx
