+++ nextjs_frontend/src/components/Header/components/NotificationFlyMenu/notification-fly-menu.module.scss
